+++ proxy/decorator.go
@@ -49,6 +49,7 @@ func NewXForwardedForDecorator() Decorator {
 func (x *xForwardedForDecorator) Decorate(req *http.Request) {
 	ip, _, _ := net.SplitHostPort(req.RemoteAddr)
 	req.Header.Set("X-Forwarded-For", ip)
+	req.Header.Set("X-Forwarded-Host", req.Host)
 	if req.TLS == nil {
 		req.Header.Set("X-Forwarded-Proto", "http")
 	} else {